A "robust" approach to testing README code
Better than a polite reminder to keep two files in sync, we could use
string-comparisons to keep them in sync.  This adds a "meta" test that
just compares the code sample in README.md to the text of
readme_test.go.  Now, if they get out of sync, that's an error.

This is me failing to take the advice of @pearkes to use Golang
examples, because I couldn't figure out how to do that in a reasonable
way.  Maybe someone cleverer than I will do it better.